+++ CHANGELOG.markdown
@@ -1,6 +1,6 @@
 3.7
 -----
-* Added `_magnitude`, `_phase` and `_conjugate` to `Data.Complex.Lens`.
+* Added `_magnitude`, `_phase` and `_conjugate` to `Data.Complex.Lens`. Renamed other lenses for consistency: `_realPart`, `_imagPart`, `_polar`.
 * Promoted `_left` and `_right` to projections and moved them to `Control.Lens.Projection`.
 * Bug fixes for `Projection`.
 * Generalized `view` and `views` to subsume the functionality of `peruse` and `peruses`.